+++ lib/dradis/plugins/projects/engine.rb
@@ -18,8 +18,8 @@ module Dradis
 
         initializer "dradis-projects.set_configs" do |app|
           options = app.config.dradis.projects
-          options.template_exporter ||= Dradis::Plugins::Projects::Export::V1::Template
-          options.template_uploader ||= Dradis::Plugins::Projects::Upload::V1::Template::Importer
+          options.template_exporter ||= Dradis::Plugins::Projects::Export::V2::Template
+          options.template_uploader ||= Dradis::Plugins::Projects::Upload::V2::Template::Importer
         end
